+++ src/locales/ar/common.json
@@ -19,5 +19,7 @@
     "confirmText": "تجاهل التغييرات",
     "message": "أنت على وشك التراجع عن كافة التغييرات التي أجريتها.",
     "title": "هل تريد تجاهل التغييرات؟"
-  }
+  },
+  "filter": "منقي",
+  "clear": "واضح"
 }
