+++ src/locales/en/components.json
@@ -89,9 +89,5 @@
       "november": "November",
       "december": "December"
     }
-  },
-  "filterPopover": {
-    "clear": "Clear",
-    "filter": "Filter"
   }
 }
