+++ src/locales/fr/common.json
@@ -19,5 +19,7 @@
     "confirmText": "Annuler les modifications",
     "message": "Vous êtes sur le point d'annuler toutes les modifications que vous avez apportées.",
     "title": "Annuler les modifications?"
-  }
+  },
+  "filter": "Filtrer",
+  "clear": "Effacer"
 }
