+++ src/locales/fr/components.json
@@ -89,9 +89,5 @@
   },
   "formField": {
     "optional": "Facultatif"
-  },
-  "filterPopover": {
-    "clear": "Effacer",
-    "filter": "Filtrer"
   }
 }
