+++ src/locales/sw/common.json
@@ -19,5 +19,7 @@
     "confirmText": "Tupa mabadiliko",
     "message": "Unakaribia kutendua mabadiliko yote uliyofanya.",
     "title": "Ungependa kutupa mabadiliko?"
-  }
+  },
+  "filter": "Chuja",
+  "clear": "Wazi"
 }
